[FLINK-13391][table] Fix InputFormatTableSource#getDataStream should not call getReturnType.
